+++ src/ENet.NET/ENetHost.cs
@@ -22,15 +22,15 @@ public class ENetHost
 {
     public Socket socket;
     public ENetAddress address; /* Internet address of the host */
-    public uint incomingBandwidth; /* downstream bandwidth of the host */
-    public uint outgoingBandwidth; /* upstream bandwidth of the host */
-    public uint bandwidthThrottleEpoch;
-    public uint mtu;
+    public long incomingBandwidth; /* downstream bandwidth of the host */
+    public long outgoingBandwidth; /* upstream bandwidth of the host */
+    public long bandwidthThrottleEpoch;
+    public long mtu;
     public uint randomSeed;
     public int recalculateBandwidthLimits;
     public ENetPeer[] peers; /* array of peers allocated for this host */
     public int peerCount; /* number of peers allocated for this host */
-    public ulong channelLimit; /* maximum number of channels allowed for connected peers */
+    public long channelLimit; /* maximum number of channels allowed for connected peers */
     public uint serviceTime;
     public ENetList<ENetPeer> dispatchQueue;
     public uint totalQueued;
@@ -45,15 +45,15 @@ public class ENetHost
     public byte packetData[2][ENets.ENET_PROTOCOL_MAXIMUM_MTU];
     public ENetAddress receivedAddress;
     public byte* receivedData;
-    public ulong receivedDataLength;
+    public long receivedDataLength;
     public uint totalSentData; /* total data sent, user should reset to 0 as needed to prevent overflow */
     public uint totalSentPackets; /* total UDP packets sent, user should reset to 0 as needed to prevent overflow */
     public uint totalReceivedData; /* total data received, user should reset to 0 as needed to prevent overflow */
     public uint totalReceivedPackets; /* total UDP packets received, user should reset to 0 as needed to prevent overflow */
     public ENetInterceptCallback intercept; /* callback the user can set to intercept received raw UDP packets */
-    public ulong connectedPeers;
-    public ulong bandwidthLimitedPeers;
-    public ulong duplicatePeers; /* optional number of allowed peers from duplicate IPs, defaults to ENET_PROTOCOL_MAXIMUM_PEER_ID */
-    public ulong maximumPacketSize; /* the maximum allowable packet size that may be sent or received on a peer */
-    public ulong maximumWaitingData; /* the maximum aggregate amount of buffer space a peer may use waiting for packets to be delivered */
+    public long connectedPeers;
+    public long bandwidthLimitedPeers;
+    public long duplicatePeers; /* optional number of allowed peers from duplicate IPs, defaults to ENET_PROTOCOL_MAXIMUM_PEER_ID */
+    public long maximumPacketSize; /* the maximum allowable packet size that may be sent or received on a peer */
+    public long maximumWaitingData; /* the maximum aggregate amount of buffer space a peer may use waiting for packets to be delivered */
 }
